fix GetOffset()
This is an invasive patch because GetOffset() is the first message to
return multiple records.  This patch changes readResponse() to take an
integer number of expected records in the message and returns [][]byte
instead of just a single []byte.  It returns an error if the wrong
number of records were read from the socket.